+++ Plutarch/Rec.hs
@@ -2,6 +2,7 @@
 
 module Plutarch.Rec (
   DataReader (DataReader, readData),
+  DataWriter (DataWriter, writeData),
   PRecord (PRecord, getRecord),
   ScottEncoded,
   ScottEncoding,
@@ -10,7 +11,10 @@ module Plutarch.Rec (
   fieldFromData,
   letrec,
   pletrec,
+  rcon,
+  recordDataFromFieldWriters,
   recordFromFieldReaders,
+  rmatch,
 ) where
 
 import Control.Monad.Trans.State.Lazy (State, evalState, get, put)
@@ -18,9 +22,21 @@ import Data.Functor.Compose (Compose (Compose, getCompose))
 import Data.Kind (Type)
 import Data.Monoid (Dual (Dual, getDual), Endo (Endo, appEndo), Sum (Sum, getSum))
 import Numeric.Natural (Natural)
-import Plutarch (PlutusType (PInner, pcon', pmatch'), pcon, phoistAcyclic, plam, plet, punsafeCoerce, (#), (:-->))
+import Plutarch (
+  PlutusType (PInner, pcon', pmatch'),
+  pcon,
+  phoistAcyclic,
+  plam,
+  plet,
+  pmatch,
+  punsafeBuiltin,
+  punsafeCoerce,
+  (#),
+  (:-->),
+ )
 import Plutarch.Bool (pif, (#==))
 import Plutarch.Builtin (PAsData, PBuiltinList, PData, pasConstr, pforgetData, pfstBuiltin, psndBuiltin)
+import Plutarch.Integer (PInteger)
 import Plutarch.Internal (
   PType,
   RawTerm (RApply, RLamAbs, RVar),
@@ -28,8 +44,9 @@ import Plutarch.Internal (
   TermResult (TermResult, getDeps, getTerm),
   mapTerm,
  )
-import Plutarch.List (phead, ptail)
+import Plutarch.List (pcons, phead, pnil, ptail)
 import Plutarch.Trace (ptraceError)
+import qualified PlutusCore as PLC
 import qualified Rank2
 
 newtype PRecord r s = PRecord {getRecord :: r (Term s)}
@@ -43,17 +60,27 @@ instance (Rank2.Distributive r, Rank2.Traversable r) => PlutusType (PRecord r) w
   pcon' :: forall s. PRecord r s -> forall t. Term s (ScottEncoding r t)
   pcon' (PRecord r) = rcon r
   pmatch' :: forall s t. (forall t. Term s (ScottEncoding r t)) -> (PRecord r s -> Term s t) -> Term s t
-  pmatch' p f = p # arg
-    where
-      arg :: Term s (ScottEncoded r t)
-      arg = Term (\i -> TermResult (RLamAbs (fieldCount (initial @r) - 1) $ rawArg i) [])
-      rawArg :: Natural -> RawTerm
-      rawArg depth = getTerm $ asRawTerm (f $ PRecord variables) $ depth + fieldCount (initial @r)
+  pmatch' p f = rmatch p (f . PRecord)
 
+-- | Convert a Haskell record value to a Scott-encoded record.
 rcon :: forall r s t. Rank2.Foldable r => r (Term s) -> Term s (ScottEncoding r t)
 rcon r = plam (\f -> punsafeCoerce $ appEndo (getDual $ Rank2.foldMap (Dual . Endo . applyField) r) f)
   where
     applyField x f = punsafeCoerce f # x
+
+-- | Match a Scott-encoded record using a function that takes a Haskell record value.
+rmatch ::
+  forall r s t.
+  (Rank2.Distributive r, Rank2.Traversable r) =>
+  (forall t. Term s (ScottEncoding r t)) ->
+  (r (Term s) -> Term s t) ->
+  Term s t
+rmatch p f = p # arg
+  where
+    arg :: Term s (ScottEncoded r t)
+    arg = Term (\i -> TermResult (RLamAbs (fieldCount (initial @r) - 1) $ rawArg i) [])
+    rawArg :: Natural -> RawTerm
+    rawArg depth = getTerm $ asRawTerm (f $ variables depth) $ depth + fieldCount (initial @r)
 
 -- | Wrapped recursive let construct, tying into knot the recursive equations specified in the record fields.
 pletrec :: forall r s. (Rank2.Distributive r, Rank2.Traversable r) => (r (Term s) -> r (Term s)) -> Term s (PRecord r)
@@ -86,35 +113,37 @@ field f = getScott (f accessors)
 
 -- | Provides a record of function terms that access each field out of a Scott-encoded record.
 accessors :: forall r s. (Rank2.Distributive r, Rank2.Traversable r) => r (ScottArgument r s)
-accessors = abstract Rank2.<$> variables
+accessors = abstract Rank2.<$> variables 0
   where
     abstract :: Term s a -> ScottArgument r s a
-    abstract (Term t) = ScottArgument (phoistAcyclic $ Term $ mapTerm (RLamAbs $ fieldCount (initial @r) - 1) . t)
+    abstract (Term t) = ScottArgument (phoistAcyclic $ Term $ mapTerm (RLamAbs $ depth - 1) . t . (depth +))
+    depth = fieldCount (initial @r)
 
 {- | A record of terms that each accesses a different variable in scope,
  outside in following the field order.
 -}
-variables :: forall r s. (Rank2.Distributive r, Rank2.Traversable r) => r (Term s)
-variables = Rank2.cotraverse var id
+variables :: forall r s. (Rank2.Distributive r, Rank2.Traversable r) => Natural -> r (Term s)
+variables baseDepth = Rank2.cotraverse var id
   where
     var :: (r (Term s) -> Term s a) -> Term s a
     var ref = ref ordered
     ordered :: r (Term s)
-    ordered = evalState (Rank2.traverse next $ initial @r) (fieldCount $ initial @r)
+    ordered = evalState (Rank2.traverse next $ initial @r) 0
     next :: f a -> State Natural (Term s a)
     next _ = do
       i <- get
-      let i' = pred i
+      let i' = succ i
       seq i' (put i')
       return $
         Term $
-          const $
+          \depth ->
             TermResult
-              { getTerm = RVar i'
+              { getTerm = RVar (depth - baseDepth - i')
               , getDeps = []
               }
 
 newtype DataReader s a = DataReader {readData :: Term s (PAsData a) -> Term s a}
+newtype DataWriter s a = DataWriter {writeData :: Term s a -> Term s (PAsData a)}
 newtype FocusFromData s a b = FocusFromData {getFocus :: Term s (PAsData a :--> PAsData b)}
 newtype FocusFromDataList s a = FocusFromDataList {getItem :: Term s (PBuiltinList PData) -> Term s (PAsData a)}
 
@@ -134,8 +163,25 @@ recordFromFieldReaders reader = DataReader $ verifySoleConstructor readRecord
     fields :: Term s (PBuiltinList PData) -> r (Compose (Term s) PAsData)
     fields bis = (\f -> Compose $ getItem f bis) Rank2.<$> fieldListFoci
 
+recordDataFromFieldWriters ::
+  forall r s.
+  (Rank2.Apply r, RecordFromData r) =>
+  r (DataWriter s) ->
+  DataWriter s (PRecord r)
+recordDataFromFieldWriters writer = DataWriter (`pmatch` writeRecord)
+  where
+    writeRecord :: PRecord r s -> Term s (PAsData (PRecord r))
+    writeRecord (PRecord r) =
+      punsafeBuiltin PLC.ConstrData # (0 :: Term s PInteger)
+        # appEndo (Rank2.foldMap (Endo . consField) (Rank2.liftA2 writeField writer r)) pnil
+    consField :: Compose (Term s) PAsData a -> Term s (PBuiltinList PData) -> Term s (PBuiltinList PData)
+    consField (Compose h) t = pcons # pforgetData h # t
+    writeField :: DataWriter s a -> Term s a -> Compose (Term s) PAsData a
+    writeField w r = Compose (writeData w r)
+
 {- | Converts a Haskell field function to a function term that extracts the 'Data' encoding of the field from the
- encoding of the whole record.
+ encoding of the whole record. If you need to access most of the record fields, it's more efficient to decode the
+ entire record at once with `recordFromFieldReaders`.
 -}
 fieldFromData ::
   RecordFromData r =>